% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_plots.R
\name{combine_plots}
\alias{combine_plots}
\title{Combining multiple plots using \code{cowplot::plot_grid()} with a title and a caption}
\usage{
combine_plots(..., title.text = NULL, caption.text = NULL,
  title.colour = "blue", caption.colour = "black", caption.size = 10,
  title.size = 16, caption.vjust = 0.5, caption.hjust = 0.5,
  title.vjust = 0.5, title.hjust = 0.5, title.rel.heights = c(0.1, 1.2),
  caption.rel.heights = c(0.1, 1.2, 0.1))
}
\arguments{
\item{...}{additional arguments that will be fed into the function \code{cowplot::plot_grid()}.}

\item{title.text}{String or plotmath expression to be drawn as title for the \emph{combined plot}.}

\item{caption.text}{String or plotmath expression to be drawn as the captionfor the \emph{combined plot}.}

\item{title.colour}{Text color for title.}

\item{caption.colour}{Text color for caption.}

\item{caption.size}{Point size of title text.}

\item{title.size}{Point size of caption text.}

\item{caption.vjust}{Vertical justification for caption. Default = 0.5 (centered on y).
0 = baseline at y, 1 = ascender at y.}

\item{caption.hjust}{Horizontal justification for caption. Default = 0.5 (centered on x).
0 = flush-left at x, 1 = flush-right.}

\item{title.vjust}{Vertical justification for title. Default = 0.5 (centered on y).
0 = baseline at y, 1 = ascender at y.}

\item{title.hjust}{Horizontal justification for title. Default = 0.5 (centered on x).
0 = flush-left at x, 1 = flush-right.}

\item{title.rel.heights}{Numerical vector of relative columns heights while combining (title, plot).}

\item{caption.rel.heights}{Numerical vector of relative columns heights while combining (title, plot, caption).}
}
\value{
Combined plot with title and caption
}
\description{
Wrapper around \code{cowplot::plot_grid()} that will return a plot grid along with a title and a caption
}
\examples{
# loading the necessary libraries
library(ggplot2)

# preparing the first plot
p1 <-
  ggplot2::ggplot(data = subset(iris, iris$Species == "setosa"),
                  aes(x = Sepal.Length, y = Sepal.Width)) + geom_point() +
  labs(title = "setosa")

# preparing the second plot
p2 <-
  ggplot2::ggplot(data = subset(iris, iris$Species == "versicolor"),
                  aes(x = Sepal.Length, y = Sepal.Width)) + geom_point() +
  labs(title = "versicolor")

# combining the plot with a title and a caption
combine_plots(
  p1,
  p2,
  labels = c("(a)", "(b)"),
  title.text = "Dataset: Iris Flower dataset",
  caption.text = "Note: Only two species of flower are displayed",
  title.colour = "red",
  caption.colour = "blue"
)

}
\author{
Indrajeet Patil
}
