% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipmisc.R
\docType{package}
\name{ipmisc}
\alias{ipmisc}
\title{\code{ipmisc}}
\description{
Collection of functions to help with certain aspects of data-wrangling and
data analysis that are not covered in the existing R packages.
}
\details{
For more documentation, see \href{https://github.com/IndrajeetPatil/ipmisc}{README} on GitHub.
}
