% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialeta_sq_ci.R
\name{partialeta_sq_ci}
\alias{partialeta_sq_ci}
\title{Confidence intervals for Partial Eta Squared}
\arguments{
\item{lm_object}{stats::lm linear model object}

\item{conf.level}{Level of confidence for the confidence interval}

\item{nboot}{Number of bootstrap samples for confidence intervals for partial eta-squared.}
}
\value{
A dataframe with results from \code{stats::lm()} with partial eta-squared and its confidence interval.
}
\description{
This function will convert a linear model object to a dataframe containing statistical details
for all effects along with partial eta-squared effect size and its confidence interval.
}
\examples{
library(datasets)
library(stats)
x <- stats::lm(formula = wt ~ mpg*cyl, data = mtcars)
ipmisc::partialeta_sq_ci(lm_object = x, conf.level = 0.95)

}
\author{
Indrajeet Patil
}
