% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_column.R
\name{signif_column}
\alias{signif_column}
\title{creating a column with significance labels}
\arguments{
\item{data}{data frame from which variables specified are preferentially to be taken}

\item{p}{the column containing p-values}
}
\value{
Returns the originally entered object (either a vector or a dataframe) in tibble format with
an additional column corresponding to statistical significance.
}
\description{
This function will add a new column to a dataframe containing p-values
}
\examples{
library(datasets)
library(stats)
library(broom)
mtcars \%>\%
stats::lm(mpg ~ wt + qsec, .) \%>\%
broom::tidy(.) \%>\%
ipmisc::signif_column(data = ., p = p.value)

}
\author{
Indrajeet Patil
}
