% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_type_switch.R
\name{stats_type_switch}
\alias{stats_type_switch}
\title{Switch type of statistics.}
\usage{
stats_type_switch(type)
}
\arguments{
\item{type}{Character string describing the type of statistics.}
}
\description{
Relevant mostly for \code{ggstatsplot} and \code{statsExpressions}
packages, where there are four types of statistics are supported:
parametric, non-parametric, robust, and Bayesian. This switch function
converts strings entered by users to a common pattern.
}
\examples{
stats_type_switch("p")
stats_type_switch("bf")
}
