% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_skim.R
\name{summary_skim}
\alias{summary_skim}
\title{custom function to get summary for multiple variables for all grouping variable levels}
\usage{
summary_skim(data, grouping.vars = ..., measures = ...)
}
\arguments{
\item{data}{dataframe}

\item{grouping.vars}{list of grouping variables}

\item{measures}{list variables for which summary needs to computed}
}
\value{
dataframe with summary statistics
}
\description{
custom function to get summary for multiple variables for all grouping variable levels
}
\examples{
library(datasets)
summary_skim(data = mtcars, grouping.vars = c(am, cyl), measures = c(wt, mpg))

}
\author{
Indrajeet Patil
}
