% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_df.R
\name{transpose_df}
\alias{transpose_df}
\title{Transposing a dataframe while preserving row and column names}
\usage{
transpose_df(df, var = "rowname", ignore.empty = FALSE)
}
\arguments{
\item{df}{Dataframe to tranpose.}

\item{var}{Name of column to use for rownames.}

\item{ignore.empty}{Default is \code{FALSE}. \code{TRUE} will ignore length-0 list elements.}
}
\examples{
data <- cbind.data.frame(x = rnorm(5), y = rnorm(5))
ipmisc::transpose_df(df = data)

}
\author{
Indrajeet Patil
}
