% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_to_wide_converter.R
\name{long_to_wide_converter}
\alias{long_to_wide_converter}
\title{Converts dataframe from long/tidy to wide format with \code{NA}s removed}
\usage{
long_to_wide_converter(
  data,
  x,
  y,
  subject.id = NULL,
  paired = TRUE,
  spread = TRUE,
  ...
)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. Other data types (e.g., matrix,table, array, etc.) will \strong{not}
be accepted.}

\item{x}{The grouping (or independent) variable from the dataframe \code{data}. In
case of a repeated measures or within-subjects design, if \code{subject.id}
argument is not available or not explicitly specified, the function assumes
that the data has already been sorted by such an id by the user and creates
an internal identifier. So if your data is \strong{not} sorted, the results
\emph{can} be inaccurate when there are more than two levels in \code{x} and there
are \code{NA}s present. The data is expected to be sorted by user in
subject-1,subject-2, ..., pattern.}

\item{y}{The response (or outcome or dependent) variable from the
dataframe \code{data}.}

\item{subject.id}{Relevant in case of a repeated measures or within-subjects
design (\code{paired = TRUE}, i.e.), it specifies the subject or repeated
measures identifier. \strong{Important}: Note that if this argument is \code{NULL}
(which is the default), the function assumes that the data has already been
sorted by such an id by the user and creates an internal identifier. So if
your data is \strong{not} sorted and you leave this argument unspecified, the
results \emph{can} be inaccurate when there are more than two levels in \code{x} and
there are \code{NA}s present.}

\item{paired}{Logical that decides whether the experimental design is
repeated measures/within-subjects or between-subjects. The default is
\code{FALSE}.}

\item{spread}{Logical that decides whether the dataframe needs to be
converted from long/tidy to wide (default: \code{TRUE}). Relevant only if
\code{paired = TRUE}.}

\item{...}{Currently ignored.}
}
\value{
A dataframe with \code{NA}s removed while respecting the
between-or-within-subjects nature of the dataset.
}
\description{
This conversion is helpful mostly for repeated measures design, where
removing \code{NA}s by participant can be a bit tedious.

It does not make sense to spread the dataframe to wide format when the
measure is not repeated, so if \code{paired = TRUE}, \code{spread} argument will be
ignored.
}
\section{Examples}{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{# for reproducibility
library(ipmisc)
set.seed(123)

# repeated measures design
long_to_wide_converter(
  data = bugs_long,
  x = condition,
  y = desire,
  subject.id = subject,
  paired = TRUE
)
}\if{html}{\out{</div>}}\preformatted{#> # A tibble: 88 x 5
#>    rowid  HDHF  HDLF  LDHF  LDLF
#>    <int> <dbl> <dbl> <dbl> <dbl>
#>  1     1  10     9     6     6  
#>  2     3  10    10    10     5  
#>  3     4   9     6     9     6  
#>  4     5   8.5   5.5   6.5   3  
#>  5     6   3     7.5   0.5   2  
#>  6     7  10    10    10    10  
#>  7     8  10     9    10    10  
#>  8     9  10     6     9.5   9.5
#>  9    11   0     0     2.5   0  
#> 10    12  10     8.5   7.5   9.5
#> # … with 78 more rows
}\if{html}{\out{<div class="sourceCode r">}}\preformatted{# independent measures design
long_to_wide_converter(
  data = ggplot2::msleep,
  x = vore,
  y = brainwt,
  paired = FALSE
)
}\if{html}{\out{</div>}}\preformatted{#> # A tibble: 51 x 3
#>    rowid vore  brainwt
#>    <int> <fct>   <dbl>
#>  1     3 carni  0.07  
#>  2     4 carni  0.0108
#>  3     5 carni  0.0256
#>  4     7 carni  0.325 
#>  5     9 carni  0.0125
#>  6    12 carni  0.157 
#>  7    17 carni  0.0175
#>  8    18 carni  0.0445
#>  9    19 carni  0.0504
#> 10    21 herbi  0.423 
#> # … with 41 more rows
}
}

