% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\alias{tibble}
\alias{enframe}
\alias{as_tibble}
\alias{\%||\%}
\alias{\%|\%}
\alias{\%<-\%}
\alias{blue}
\alias{green}
\alias{red}
\alias{yellow}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{crayon}{\code{\link[crayon]{blue}}, \code{\link[crayon]{green}}, \code{\link[crayon]{red}}, \code{\link[crayon]{yellow}}}

  \item{magrittr}{\code{\link[magrittr]{\%$\%}}, \code{\link[magrittr]{\%<>\%}}, \code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{\%|\%}}, \code{\link[rlang]{\%||\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{enframe}}, \code{\link[tibble]{tibble}}}

  \item{zeallot}{\code{\link[zeallot]{\%<-\%}}}
}}

