% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_column.R
\name{signif_column}
\alias{signif_column}
\title{creating a column with significance labels}
\usage{
signif_column(data = NULL, p)
}
\arguments{
\item{data}{data frame from which variables specified are preferentially to be taken}

\item{p}{the column containing p-values}
}
\description{
This function will add a new column to a dataframe containing p-values
}
\examples{
# the most common error is not adding quotes to column names that contain spaces
signif_column(data = data, p = `p-value (adjusted)`)
signif_column(data = data, p = `Pr(>|t|)`)
signif_column(data = data$p.value)

}
\author{
Indrajeet Patil
}
