% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_type_switch.R
\name{stats_type_switch}
\alias{stats_type_switch}
\title{Switch type of statistics.}
\usage{
stats_type_switch(type)
}
\arguments{
\item{type}{A character specifying the type of statistical approach:
\itemize{
\item \code{"parametric"}
\item \code{"nonparametric"}
\item \code{"robust"}
\item \code{"bayes"}
}

This argument also accepts the following abbreviations: \code{"p"} (for
\emph{parametric}), \code{"np"} (for \emph{nonparametric}), \code{"r"} (for \emph{robust}), \code{"bf"}
(for \emph{Bayes Factor} or \emph{Bayesian}).}
}
\description{
Relevant mostly for \code{ggstatsplot} and \code{statsExpressions} packages, where
different statistical approaches are supported via this argument: parametric,
non-parametric, robust, and Bayesian. This switch function converts strings
entered by users to a common pattern for convenience.
}
\examples{
stats_type_switch("p")
stats_type_switch("bf")
}
