% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_type_switch.R
\name{stats_type_switch}
\alias{stats_type_switch}
\title{Switch type of statistics.}
\usage{
stats_type_switch(type)
}
\arguments{
\item{type}{A character specifying the type of statistical approach:
\itemize{
\item \code{"parametric"}
\item \code{"nonparametric"}
\item \code{"robust"}
\item \code{"bayes"}
}}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}

Relevant mostly for \code{ggstatsplot} and \code{statsExpressions} packages, where
different statistical approaches are supported via this argument: parametric,
non-parametric, robust, and Bayesian. This switch function converts strings
entered by users to a common pattern for convenience.
}
\examples{
stats_type_switch("p")
stats_type_switch("bf")
}
