% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_skim.R
\name{summary_skim}
\alias{summary_skim}
\title{Custom function to get descriptive statistics for multiple numeric variables for all grouping variable levels}
\usage{
summary_skim(data, grouping.vars, measures)
}
\arguments{
\item{data}{dataframe}

\item{grouping.vars}{list of grouping variables}

\item{measures}{list variables for which summary needs to computed (only \emph{numeric} variables should be entered)}
}
\value{
Dataframe with descriptive statistics for numeric variables (n, mean, sd, median, min, max)
}
\description{
Custom function to get descriptive statistics for multiple numeric variables for all grouping variable levels
}
\examples{
library(datasets)
# if you have multiple variable for each argument
summary_skim(data = mtcars, grouping.vars = c(am, cyl), measures = c(wt, mpg))
# if you have just one variable per argument
summary_skim(data = mtcars, grouping.vars = am, measures = wt)
# or
summary_skim(data = mtcars, grouping.vars = c(am), measures = c(wt))

}
\author{
Indrajeet Patil
}
