% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_df.R
\name{transpose_df}
\alias{transpose_df}
\title{Transposing a dataframe while preserving row and column names}
\usage{
transpose_df(df, var = "rowname", ignore.empty = FALSE)
}
\arguments{
\item{df}{Dataframe to tranpose.}

\item{var}{Name of column to use for rownames.}

\item{ignore.empty}{Default is \code{FALSE}. \code{TRUE} will ignore length-0 list elements.}
}
\value{
A transposed dataframes with row and column names.
}
\description{
Although functions like \code{data.table::transpose()} or \code{base::t()} can transpose dataframes well, they
don't preserve row and column names. The current function transposes dataframes while preserving row and column names.
}
\examples{
set.seed(402)
data <- cbind.data.frame(x = rnorm(5), y = rnorm(5))
transpose_df(df = data)

}
\author{
Indrajeet Patil
}
